ai-service/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.4
distro==1.9.0
fastapi==0.128.0
filetype==1.2.0
google-auth==2.48.0
google-genai==1.62.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==1.2.11
langchain-google-genai==4.2.0
langchain-openai==1.1.9
langgraph==1.0.8
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.5
langsmith==0.7.1
openai==2.17.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==25.0
pip==25.3
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
setuptools==80.10.2
sniffio==1.3.1
starlette==0.50.0
tenacity==9.1.4
tiktoken==0.12.0
tqdm==4.67.3
typing_extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
uuid_utils==0.14.0
uvicorn==0.40.0
websockets==15.0.1
wheel==0.46.3
xxhash==3.6.0
zstandard==0.25.0
